requirements.txt
dlib
numpy
opencv-python
scikit-learn
scipy
joblib
pathlib
more-itertools
deepface
tensorflow
Pillow